Add files via upload
This Excel Project is about Excel Data Modelling, Analysis and Visualisation. The project is about a pivot table (and a pivot chart) to show the sales by product and region from a US superstore specialising in office furniture and equipment. The user will be able to filter by order date, and slice by Customer Segment, Product Category and Product Sub-Category